Struct nsl_duplex_d2::DuplexD2

pub struct DuplexD2 { /* fields omitted */ }

Structure for interacting with Duplex-D2 Radio API

Methods

impl DuplexD2

pub fn new(conn: Connection) -> DuplexD2

Constructor for DuplexD2 structure

pub fn get_state_of_health_for_modem(&self) -> RadioResult<StateOfHealth>

Retrieves a record of information regarding modem functioning.

pub fn get_uploaded_file_count(&self) -> RadioResult<u32>

Retrieves a count of files that have been received by the modem and await retrieval by the FP.

pub fn get_uploaded_message_count(&self) -> RadioResult<u32>

Retrieves a count of messages that have been received by the modem and await retrieval by the FP.

pub fn get_download_file_count(&self) -> RadioResult<u32>

Retrieves a count of messages that have been received by the modem and await retrieval by the FP.

pub fn get_geolocation_position_estimate(&self) -> RadioResult<GeoRecord>

Retrieves an estimate of the modem’s latitude and longitude coordinates at the time of the last connection.

pub fn get_uploaded_file(&self) -> RadioResult<File>

Retrieves the next file in the upload queue. File is then ACKed and deleted from queue.

pub fn get_uploaded_message(&self) -> RadioResult<File>

Retrieves the next message in the upload queue. Message is then ACKed and deleted.

pub fn delete_download_files(&self) -> RadioResult<u32>

Deletes all files in the modem download queue. Returns number of files deleted.

pub fn delete_uploaded_files(&self) -> RadioResult<u32>

Deletes all files in the modem upload queue. Returns number of files deleted.

pub fn delete_uploaded_messages(&self) -> RadioResult<u32>

Deletes all messages in the modem upload queue. Returns number of messages deleted.

pub fn put_download_file(&self, file: &File) -> RadioResult<bool>

Put a new file on the download queue, returns true if accepted by EyeStar-D2.

pub fn get_alive(&self) -> RadioResult<bool>

Check if modem is powered and if it is able to respond to commands.